Merged login&signup, css update, added JS and imgs
Old login and signup htmls were removed. Added a
new html that is the merged version of the login
and signup. Images were added to accomodate our
github page. CSS file is significantly modified to
account for new design. Javascript file is added,
**HOWEVER**, it is not fully complete. May have
missed some edge cases and need to accomodate the
JSON format of the API. Comments were added for
areas that I noticed that needed to be changed.