Added Edit Contact Popup Form
Duplicated the Add Contact popup form to also be used as a Edit
Contact form. Classes and IDs for the Edit Contact form were edited
with the old names remaining the same. However, they should also be
changed to allow for precision and cleaner code.

ie. popupForm should be renamed to addPopupForm