Edit userID names and modified JSON format to test
Modified the capitalization of the userID key in the JSON object for
compatibility. Added ID as a needed key from the searchContact. Still
have not made a solution to having the checkbox appear for the
signupPassword when criteria is met. Possible change is removing the
buttons from the delete contact row.