Rearranged some lines (editContact) for serv. test
Having a issue where the edit contact form was not showing up. The
debugger was saying that one of the global variable had an issue with
TDZ so placing the variable reassignment after making the form show up.
Deleted some old testing data that were no longer being used. Placed the
global variables to the top of the start for the contact page javascript