Changed some global variables to var (counter TDZ)
Getting an issue with the editContact function where the global
variables (originally used let contactDataToEdit = null;) kept
giving a reference error. It seemed to be an issue with the TDZ
where the variables are not hoisted. So now attempting to use
var (know it is a bad practice) for the global variables instead
to see if the TDZ issue persists.