Minor additions to make functions more seemless
Some minor changes were made to make transitions more smooth.
Such as successfully editing a contact will close the edit
contact form and display the result under the search bar instead.
Also implemented a function to clear the table if no result for
searched term exist but unsure if it is correctly done.