Fixed Enter keypress, added doSignin for doSignup
Enter keypress was not correctly implemented, added possible
fix for event listener. Duplicated the doLogin function and modified
it for the signup function. This is so that if the account creation
was successful, it will directly log you in without prompting you to
go back to the login screen to re-enter the credentials.