Minor fixes to carousel, still needs testing
Known bug: deleting a contact causes the delete contact form to
pop up twice, requiring user to click twice to close the delete form.
Tried to call the closeDeleteForm twice to see if it closes the
duplicate form.

Tried to fix: clicking the search while on carousel form causes the
pagination buttons to appear.